Fix GitHub issue #2.
It was a problem when a group was removed but some of its members were
not, and yet they expected to belong to it.  This could happen either
by an error in the profile, or because a group was removed but some of
its members was part of the kept_users dictionary.