sanitize filenames
A name with spaces or slashes would be used 'as-is' for logfiles and
this would cause a resultant command to fail
(e.g. a crontab entry of 'do-something >log file').  With this
patch, '/' and whitespace characters are transformed to
underscores. People can of course still use the 'log' attribute
within the cron definition to specify a different name, but the
default will no longer cause an error.

This patch is in reply to support request #1 ("cron component bug")
in sourceforge.