Remove spurious check for a NoAction flag.
`--noaction` in this case is handled directly by the CAF file handles.

Blocking the execution of commit_configuration prevents the admin from
seeing the diffs of any files.

CAF::File* guarantee nothing will be changed in the system.

Of course, this assumes the user's perl-CAF and ncm-ncd are less than 3
years old.