Drop dependencies on IPS from the RPM
The IPS work contains

use IPS;

statements, which by default rpmbuild translates into dependencies on
perl(IPS).  That code is solaris-specific, but we don't want to depend
on it.

The alternative would be to remove NCM::Component::spma::ips from the
RPM, but it's terribly cumbersome to do it with Maven.