config.pan should include the common config
It was observed that config-common was only being included by
config-xml.pan and not config.pan or config-rpm.pan which led to a
regression when upgrading components from before the xml/rpm split was
introduced.

The fix is to make config.pan always include config-common.pan. The
templates are unique so including them twice should not cause any
problems.