Fix bug in smear calculations and added test
If a cron item (not a list) is to be smeared, and is the biggest item allowed
in the range (59m, Sat, etc) it is incorrectly rounded to the next
value. For example this causes anything that is to be smeared on a
Saturday to be set to a Sunday.

Also add a test to exercise the smear bug.