ncm-sysctl: validate that confFile ends in .conf
systemd requires that config files in /etc/sysctl.d end in .conf. It is
not required for non-systemd OSes, but enforcing the constraint now
makes it easier for nodes to be upgraded to systemd.

See: http://www.freedesktop.org/software/systemd/man/sysctl.d.html

Also validate that the config file path is either an absolute path, or a
single filename since relative paths are not valid.