Fix a cron race condition bug
There is a race condition: when the component mass creates files (within a
second) cron may read in just some of the cronjobs ignoring all the others
until crond restarted or the host is rebooted. RedHat 5 has this bug.
RH6 uses inotify and not suspectible to this.

By touching /etc/cron.d we force crond to re-read all the files (cronjobs).