Remove code breaking downgrade on RHEL5.
The lack of package protection support in RHEL5's YUM can cause
essential libraries and packages to be removed upon downgrade
irreversibly breakng the system. The downgrade should work
flawlessly for all RHELs if this code is removed. The next code
block implementing the functionality properly works correctly on
all RHELs so this removed code block was superfluous and can
safely be removed.