ncm-grub: support setting password in grub.conf.
It is possible to set the password in the profile, or alternatively
refer to a file which the component will use to look it up. The file
should have the user (default is root) and then password in colon
delimited fields (like /etc/shadow).

While doing this, setting the serial and terminal settings has been
refactored to use CAF::FileEditor.