ncm-download: fix proxy handling.
Bug in the proxy handling. The list of proxies is passed to a function
by reference, which is intended to rotate the list.

While the intention (and previous behaviour) was to put $proxy back
into the original list, either at the front or at the end, now it creates
a new list instead while the original one gets shorter...