ncm-grub: support setting encrypted password for grub2.
If /boot/grub2 exists, set GRUB2_PASSWORD in /boot/grub2/user.cfg.

In el7 /etc/grub.d/01_users is part of grub2-tools for both UEFI and
BIOS bootloaders, it checks /boot/grub2/user.cfg for a GRUB2_PASSWORD
and configures it as a root superuser. So regenerating the grub config
is not necessary.