Add support for pkg exact-install and whitelists
Where it is possible to use the `pkg exact-install` command to create a
new BE with an exact list of packages, then use it as it is faster.
Where it is not possible to do this, revert to the previous logic where
the package list is determined by comparing with a temporary package
image directory.

This patch also adds support for whitelists which is a list of packages
that are permitted on the system even if they have not been explicitly
requested and even if `userpkgs` is set to `no`.