ncm-nfs: stop nfs reload when adding comments to empty exports file
On ncm-nfs's first run, it adds comment lines to the exports file. Even If there are no exports this will trigger a reload of the nfs service. On RHEL6 this appears harmless as the reload returns 0, but on RHEL 7, systemd returns an error:
<code>
# /bin/systemctl reload nfs; echo $?
Job for nfs-server.service invalid.
1
</code>
 
This patch attempts to stop this. I think should partially address #1029 but any changes to fstab may also make the component reload nfs