ncm-systemd: Perform actions one unit at a time
If performing an action fails for one unit (e.g., because one tries to
enable a service which is not installed), then systemctl usually returns
a non-descriptive error like "No such file or directory", without any
indications for which unit caused the failure. What's worse, the rest of
the units will not be acted on either, often resulting in much more
breakage than the problematic unit would warrant.

To avoid this, perform actions one unit at a time.