metaconfig: Update xinetd schema to support more configs
"port" added as a supported option.

"instances" is now a string to allow the setting of all the valid
values. Previously it was docuented that leaving it unset gave
UNLIMITED, however the actual behaviour is to inherit from
/etc/xinetd.conf. On RHEL the default value is 50 so the schema must
allow this to be set. With this change you can set "UNLIMITED" or a
quoted number and the right thing will happen in the service file.

"cps" removed default setting.
"per_source" removed default setting.
These two changes mean services will take their defaults from
/etc/xinetd.conf which is more obvious behaviour.