Dockerfile: adjust user directive and maven env
There is some test code which calls getpwuid($<) to find the current
username. If this returns undef then a Perl warning is generated which
fails the tests. Workaround this by using uid=99 which maps to user
nobody in the base image, but cannot write to its $HOME. Maven must
therefore be configured to use an temporary directory instead of $HOME.