ncm-hostsfile: Test IPv6 and dual stack localhosts
In these we show how dual stack entries can set takeover and overwrite the existing localhost entry or not,
specifying aliases for both stacks that allow `localhost` to resolve for both protocols while still avoiding duplicate entries.

As with other components we also split config into config-common and config-rpm to allow schema and config to be used in tests.