ncm-spma: apt: Warn if any packages appear to be unavailable
This could be because they have been renamed (rare) or they are virtual
(common) in both cases when run interactively apt will warn the user:

e.g.
> Note, selecting 'man-db' instead of 'man'

However this warning is supressed when the output is not a TTY so we
have to contrive another means of identifying these cases.