ncm-metaconfig: rsyslog: Add support for conditional actions
The rsyslog module previously could not use conditional filters for
every type of action, this change updates the templates and the schema
to add a new "condition" field to the generic rsyslog action template.
Whatever is put in the condition field will be inserted as the condition
of an if block surrounding the action, which is generated using the
existing action template.

This means that the "condition" field can now be added to any action
with the generic rsyslog action template. This has also been added to the
usermessage action which incorrectly claimed that it would by default
only output messages with the emerg priority. This behaviour can
actually only be acheived by surrounding the statement with a condition.
The confusion was probably caused by a misleadingly labeled example in
the rsyslog documentation for omusrmsg.