ncm-metaconfig: rsyslog: Add generic condition support
The rsyslog module previously could not use conditional filters for
every type of action, this change updates the templates and the schema
to add a new "condition" field to the generic rsyslog action template.
Whatever is put in the condition field will be inserted as the condition
of an if block surrounding the action, which is generated using the
existing action template.

This means that the "condition" field can now be added to any action.